move io utils to topos/io
