+++ topos/__init__.py
@@ -1,5 +1,6 @@
 from .core import *
-import topos.base as base
+import topos.base   as base
+import topos.io     as io
 
 #old
 #from .domain import *
