+++ topos/base/__init__.py
@@ -1,5 +1,4 @@
 from .graph         import Graph
-from .shape         import Shape
 # old
 #from .set           import Set, Iterable, Map, Mapping
 #from .fiber         import Fiber
