+++ topos/io/__init__.py
@@ -1,0 +1,1 @@
+from .io_utils import alignString, parseTensor
